refactor: 사다리 생성 하는 로직 Ladder 클래스로 이동, 새로운 객체 얻을 때 사용하도록 변경
