+++ src/main/java/ladder/domain/Ladder.java
@@ -1,5 +1,8 @@
 package ladder.domain;
 
+import ladder.util.RandomHelper;
+
+import java.util.ArrayList;
 import java.util.List;
 
 public class Ladder {
@@ -10,7 +13,20 @@ public class Ladder {
     }
 
     public static Ladder newLadder(LadderGameData ladderGameData) {
-        return new Ladder(LadderMaker.generateLadder(ladderGameData));
+        List<Line> lines = new ArrayList<>();
+        for (int i = 0; i < ladderGameData.getHeight(); i++) {
+            lines.add(new Line(generateOneLine(ladderGameData.getPerson().getCountOfPerson())));
+        }
+        return new Ladder(lines);
+    }
+
+    private static List<Boolean> generateOneLine(int countOfPerson) {
+        List<Boolean> points = new ArrayList<>();
+        for (int j = 0; j < countOfPerson; j++) {
+            points.add(RandomHelper.randomPoint(points, countOfPerson));
+        }
+        points.add(false);
+        return points;
     }
 
     int moveStartToEnd(int index) {
